WIP: streamlining the manipulators in native floating-point
