+++ include/universal/native/nonconstexpr/msvc_long_double.hpp
@@ -19,6 +19,10 @@ inline std::tuple<bool, int, std::uint64_t> ieee_components(long double fp) {
 	return ieee_components(double(fp));
 }
 
+inline std::string to_hex(long double fp) {
+	return to_hex(double(fp));
+}
+
 // specialization for IEEE long double precision floats
 inline std::string to_base2_scientific(long double number) {
 	return to_base2_scientific(double(number));
